requirements.txt
check-jsonschema==0.35.0
pytest==8.4.2
pytest-cov==6.0.0
pytest-timeout==2.3.1
PyYAML==6.0.2
ruff==0.13.0
pandas==2.3.3
tabulate==0.9.0
streamlit>=1.28.0
